Simplify mouse object tracker
I've given up on pixel perfect mouse hit detection. It might be possible with a component designed specifically for it, but that still wouldn't work well with rotated objects and it wouldn't work at all with oddly shaped ones. The minor error is a pretty small price to pay to just use Unity's built in collision detection, I guess.